[jwbron/eval-measurement-tool] review: split live-ab report shapes/renderers into live-ab-report.ts
CI's max-lines (1000) caught live-ab.ts at 1011 after the repeats and
arbiter work; this worktree cannot run eslint locally (dot-dir ignore).
The report types (ArmRunReport, AbReport, MultiAbReport, gate types) and
both markdown renderers move to a leaf module with no runner dependency;
live-ab.ts re-exports them so the import surface is unchanged.
